chore: implement submitted and annual report POMs
