+++ bciers/apps/reporting-e2e/tests/workflows/LFO/view-report.spec.ts
@@ -3,6 +3,8 @@ import { UserRole } from "@bciers/e2e/utils/enums";
 import { FacilityIDs, OPERATION_NAMES } from "@/reporting-e2e/utils/enums";
 import { CurrentReportsPOM } from "@/reporting-e2e/poms/current-reports";
 import { takeStabilizedScreenshot } from "@bciers/e2e/utils/helpers";
+import { SubmittedPOM } from "@/reporting-e2e/poms/submitted";
+import { AnnualReportPOM } from "@/reporting-e2e/poms/annual-report";
 
 const test = setupBeforeAllTest(UserRole.INDUSTRY_USER_ADMIN);
 const internalTest = setupBeforeAllTest(UserRole.CAS_ANALYST);
@@ -14,7 +16,6 @@ test.describe("LFO: view a submitted report for the current reporting year", () 
     page,
     happoScreenshot,
   }) => {
-    const isExternalUser = true;
     const isLFO = true;
     // ── 1. Navigate to the past reports grid ──
     const grid = new CurrentReportsPOM(page);
@@ -25,22 +26,22 @@ test.describe("LFO: view a submitted report for the current reporting year", () 
     // ── 2. Click "View Details" for the report — navigates to the 'Submitted' report view ──
     await grid.reportHistoryForOperation(OPERATION_NAMES.BANANA_LFO);
     await grid.viewDetailsFromReportHistory();
-    await grid.verifySubmittedReportView(
+
+    const submittedReport = new SubmittedPOM(page);
+    await submittedReport.verifySubmittedReportView(
       OPERATION_NAMES.BANANA_LFO,
       isLFO,
-      isExternalUser,
     );
 
     await takeStabilizedScreenshot(happoScreenshot, page, {
       component: "LFO Report - Submitted",
       variant: "external",
     });
-    await grid.viewDetailsFromFacilityGrid(
+    await submittedReport.viewDetailsFromFacilityGrid(
       "Facility 1",
       FacilityIDs.BANANA_LFO_1,
-      isExternalUser,
     );
-    await grid.verifyFacilitySubmittedReportView("Facility 1");
+    await submittedReport.verifyFacilitySubmittedReportView("Facility 1");
     await takeStabilizedScreenshot(happoScreenshot, page, {
       component: "LFO Facility Report - Submitted",
       variant: "external",
@@ -67,10 +68,11 @@ test.describe("LFO: view a submitted report for the current reporting year", () 
             isExternalUser,
           );
 
-          await grid.verifySubmittedReportView(
+          const submittedReport = new AnnualReportPOM(page);
+
+          await submittedReport.verifySubmittedReportView(
             OPERATION_NAMES.BANANA_LFO,
             isLFO,
-            isExternalUser,
           );
 
           await takeStabilizedScreenshot(happoScreenshot, page, {
@@ -79,12 +81,11 @@ test.describe("LFO: view a submitted report for the current reporting year", () 
           });
 
           // -- 3. Click "View Details" for the facility report — navigates to the facility 'Submitted' report view ──
-          await grid.viewDetailsFromFacilityGrid(
+          await submittedReport.viewDetailsFromFacilityGrid(
             "Facility 1",
             FacilityIDs.BANANA_LFO_1,
-            isExternalUser,
           );
-          await grid.verifyFacilitySubmittedReportView("Facility 1");
+          await submittedReport.verifyFacilitySubmittedReportView("Facility 1");
           await takeStabilizedScreenshot(happoScreenshot, page, {
             component: "LFO Facility Report - Submitted",
             variant: "internal",
