+++ bciers/apps/reporting-e2e/utils/helpers.ts
@@ -10,3 +10,17 @@ export async function verifyFormTitle(
     ),
   ).toHaveText(title);
 }
+
+export async function verifySaveAsPDF(page: Page): Promise<void> {
+  const saveAsPDFButton = page.getByRole("button", {
+    name: /Save as PDF/i,
+  });
+  await expect(saveAsPDFButton).toBeVisible();
+  await expect(saveAsPDFButton).toBeEnabled();
+
+  await page.evaluate(
+    "(() => {window.waitForPrintDialog = new Promise(f => window.print = f);})()",
+  );
+  await saveAsPDFButton.click();
+  await page.waitForFunction("window.waitForPrintDialog");
+}
